+++ bin/web.dart
@@ -1,5 +1,26 @@
+import 'package:args/args.dart';
 import 'package:pubspec_checker/pubspec_checker.dart';
 
 Future<void> main(List<String> arguments) async {
-  await PackageChecker().checkPlatform(PackagePlatform.web);
+  final parser = ArgParser()
+    ..addFlag('show',
+        abbr: 's',
+        negatable: false,
+        help: 'Show list of packages with supported platforms')
+    ..addFlag('links',
+        abbr: 'l', negatable: false, help: 'Show links of packages');
+
+  // Parse the arguments
+  final results = parser.parse(arguments);
+
+  // Extract platforms to be checked (positional arguments)
+  final showList = results['show'] as bool; // Check if -s is provided
+  final showLinks = results['links'] as bool; // Check if -l is provided
+
+  if (showList) {
+    await PackageChecker()
+        .checkAll([PackagePlatform.web.name], showLink: showLinks);
+  } else {
+    await PackageChecker().check(PackagePlatform.web.name, showLink: showLinks);
+  }
 }