+++ lib/src/package_platform.dart
@@ -8,6 +8,7 @@ enum PackagePlatform { android, ios, windows, macos, linux, web }
 List<Map<String, dynamic>> platforms = [
   {'id': PackagePlatform.android, 'name': 'android'},
   {'id': PackagePlatform.ios, 'name': 'ios'},
+  {'id': PackagePlatform.macos, 'name': 'macos'},
   {'id': PackagePlatform.windows, 'name': 'windows'},
   {'id': PackagePlatform.linux, 'name': 'linux'},
   {'id': PackagePlatform.web, 'name': 'web'},